pyproject.toml
[project]
name = "adk-postcards"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[mcp]>=1.27.0",
]
